nb - renamed Messages folder to ChannelMessages to reflect the naming conventions in the story
